refactor: move announcement tasks to own file
